pyproject.toml
[project]
name = "plab-rundown"
version = "0.1.0"
description = "The Rundown AI 邮件翻译与公众号发布系统"
authors = [
    {name = "Your Name", email = "your.email@example.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    # Web 框架
    "fastapi>=0.104.1",
    "uvicorn[standard]>=0.24.0",
    "python-dotenv>=1.0.0",
    
    # Gmail API
    "google-auth>=2.23.4",
    "google-auth-oauthlib>=1.1.0",
    "google-auth-httplib2>=0.1.1",
    "google-api-python-client>=2.108.0",
    
    # AI 和 LangChain
    "langchain>=0.1.0",
    "langchain-openai>=0.0.2",
    "langchain-google-vertexai>=0.1.0",
    "langchain-google-genai>=0.0.5",
    "openai>=1.3.7",
    "google-cloud-aiplatform>=1.38.0",
    "PySocks>=1.7.1",
    
    # HTTP 请求
    "requests>=2.31.0",
    "httpx[socks]>=0.25.2",
    
    # HTML 解析
    "beautifulsoup4>=4.12.2",
    "lxml>=4.9.3",
    "html2text>=2020.1.16",
    
    # 数据库 - PostgreSQL (Supabase)
    "sqlalchemy>=2.0.23",
    "psycopg2-binary>=2.9.9",
    "alembic>=1.13.0",
    
    # 任务调度
    "apscheduler>=3.10.4",
    
    # 日期时间
    "python-dateutil>=2.8.2",
    "pytz>=2023.3",
    
    # 工具
    "pydantic>=2.5.2",
    "pydantic-settings>=2.1.0",
    "pyyaml>=6.0.1",
]

[project.optional-dependencies]
dev = [
    # 测试
    "pytest>=7.4.3",
    "pytest-asyncio>=0.21.1",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.12.0",
    
    # 代码质量
    "black>=23.12.0",
    "flake8>=6.1.0",
    "mypy>=1.7.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src"]

[tool.black]
line-length = 100
target-version = ['py311']

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = "test_*.py"
python_classes = "Test*"
python_functions = "test_*"
addopts = "-v --tb=short"

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

